create centralized lattice.py for shared functions
